refactor: polls design
